Rename modelica_lang => modelicalang, ModelicaLanguage => ModelicaLang
