+++ document/index.rst
@@ -1,9 +1,9 @@
-.. ModelicaLanguage documentation master file, created by
+.. ModelicaLang documentation master file, created by
    sphinx-quickstart on Mon Jan 30 11:20:24 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to ModelicaLanguage's documentation!
+Welcome to ModelicaLang's documentation!
 ============================================
 
 .. toctree::
